+++ js/primitivenotes.js
@@ -6,7 +6,7 @@
  * @copyright Copyright (c) 2022, Offerel
  * @license GNU General Public License, version 3
  */
-var mde, tagify;
+var mde, tagify, oNote;
 var filelist = [];
 var loader = document.createElement("div");
 var ldr = document.createElement("div");
@@ -157,6 +157,19 @@ window.rcmail && rcmail.addEventListener("init", function(a) {
 			manageDropUpload(e.dataTransfer.files);
 		});
 	}
+	
+	window.addEventListener("keydown", function(e) {
+		if(e.shiftKey && e.ctrlKey && "KeyF" === e.code) {
+			document.getElementById("notessearchform").focus();
+			return false;
+		}
+		
+		if(e.ctrlKey && "KeyF" === e.code) {
+			e.preventDefault();
+			searchInNote();
+			return false;
+		}
+	});
 
 	document.addEventListener("keyup", event => {
 		if(event.key == 'Escape') {
@@ -168,10 +181,6 @@ window.rcmail && rcmail.addEventListener("init", function(a) {
 				document.getElementById("notessearchform").value = '';
 				document.getElementById("notessearchform").dispatchEvent(new KeyboardEvent('keyup', {'key':''}));
 			}
-		}
-
-		if(event.code == 'KeyF' && (event.ctrlKey || event.metaKey)) {
-			document.getElementById("notessearchform").focus();
 		}
 	});
 	document.getElementById('source').addEventListener('click', osource, true);
@@ -197,6 +206,127 @@ window.rcmail && rcmail.addEventListener("init", function(a) {
 		rcmail.http_post('displayNote', postData, false);
 	}
 });
+
+function searchInNote() {
+	let cSearch = document.createElement('div');
+	cSearch.id = 'cSearch';
+	let iSearch = document.createElement('input');
+	iSearch.id = 'iSearch';
+	cSearch.appendChild(iSearch);
+	let iInfo = document.createElement('div');
+	let act = document.createElement('span');
+	act.innerText = '0';
+	let found = document.createElement('span');
+	
+	let prev = document.createElement('a');
+	prev.id = 'prev';
+	prev.innerHTML = '&#10092;'
+	prev.addEventListener('click', e => {
+		e.preventDefault();
+		let fpos = parseInt(act.innerText) - 1;
+		fpos = (fpos >= 1) ? fpos:1;
+		document.querySelector('[data-pos="' + fpos.toString() + '"]').scrollIntoView();
+		document.querySelectorAll('mark').forEach(e => {
+			e.classList.remove('fc')
+		});
+		document.querySelector('[data-pos="' + fpos.toString() + '"]').classList.add('fc');
+
+		if(fpos == 1) {
+			prev.classList.add('anone');
+			next.classList.remove('anone');
+		} else {
+			prev.classList.remove('anone');
+			next.classList.remove('anone');
+		}
+
+		act.innerText = fpos;
+	});
+	
+	let next = document.createElement('a');
+	next.id = 'next';
+	next.innerHTML = '&#10093;'
+	next.addEventListener('click', e => {
+		e.preventDefault();
+		marks = document.querySelectorAll('mark');
+		let fpos = parseInt(act.innerText) + 1;
+		fpos = (fpos > marks.length) ? marks.length:fpos;
+		document.querySelector('[data-pos="' + fpos.toString() + '"]').scrollIntoView();
+		marks.forEach(e => {
+			e.classList.remove('fc')
+		});
+		document.querySelector('[data-pos="' + fpos.toString() + '"]').classList.add('fc');
+
+		if(fpos == marks.length) {
+			console.log(fpos);
+			next.classList.add('anone');
+			prev.classList.remove('anone');
+		} else {
+			next.classList.remove('anone');
+			prev.classList.remove('anone');
+		}
+
+		act.innerText = fpos;
+	});
+	
+	let iStatus = document.createElement('span');
+	let iText = document.createTextNode(' / ');
+	
+	found.innerText = '0';
+	let iText2 = document.createElement('span');
+	iText2.classList.add('sep');
+	
+	iStatus.appendChild(act);
+	iStatus.appendChild(iText);
+	iStatus.appendChild(found);
+	iStatus.appendChild(iText2);
+	
+	iInfo.appendChild(iStatus);
+	iInfo.appendChild(prev);
+	iInfo.appendChild(next);
+	
+	cSearch.appendChild(iInfo);
+	let main_area = document.getElementById('main_area');
+	main_area.appendChild(cSearch);
+	iSearch.focus();
+	iSearch.addEventListener("keyup", e => {
+		if(e.key == 'Escape') {
+			cSearch.remove();
+			iSearch.remove();
+		} else {
+			let sPos = 1;
+			let sTerm = document.getElementById('iSearch').value.trim();
+			let text = oNote;
+			let re = new RegExp(sTerm,"gi");
+			let newText = '';
+			let marks;
+			
+			if(sTerm.length < 3) {
+				newText = oNote;
+			} else {
+				newText = oNote.replace(re, `<mark>${sTerm}</mark>`);
+			}
+			
+			mde.value(newText);
+			
+			marks = document.querySelectorAll('mark');
+			marks.forEach(e => {
+				e.dataset.pos = sPos;
+				sPos++;
+			});
+			
+			found.innerText = (marks.length) ? marks.length:0;
+			
+			try {
+				document.querySelector('[data-pos="1"]').scrollIntoView(true);
+				act.innerText = '1';
+				prev.classList.add('anone');
+				document.querySelector('[data-pos="1"]').classList.add('fc')
+			} catch (e) {
+				return null;
+			}
+		}
+	});
+}
 
 function uplMedia() {
 	document.getElementById("dropMedia").click();
@@ -555,6 +685,7 @@ function loadNote(response) {
 	document.getElementById(response.note.id).classList.add('lselected');
 
 	if(response.note.mime_type.indexOf('text') === 0) {
+		oNote = response.note.content;
 		mde.value(response.note.content);
 		document.querySelector('.EasyMDEContainer').classList.remove('mdeHide');
 		setTimeout(() => {