+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### v2.1.3
+- Added search function
 - Fix auto-copy code
 - Fix detect paste data
 - Fix for notes folder not available